model-server/requirements.txt
Flask
tensorflow
Pillow
numpy
google-cloud-storage==2.10.0
google-cloud-core==2.3.3
google-auth
google-auth-oauthlib
gunicorn 